refactor(px): use request-response instead of dialing back
